fourier order int wasn't pulled in
Not sure why (I _thought_ I pulled all recent commits...).

Former-commit-id: 2478d5235e40586477725a2dd311e9d34cbe41f5

+++ astrobase/varbase/lcfit.py
@@ -236,7 +236,7 @@ def fourier_fit_magseries(times, mags, errs, period,
     # with respect to phase -- the light curve minimum)
     ptimes = stimes[phasesortind]
 
-    fourierorder = len(initfourierparams)/2
+    fourierorder = int(len(initfourierparams)/2)
 
     LOGINFO('fitting Fourier series of order %s to '
             'mag series with %s observations, '
